Renaming onnx.py to onnx_reader.py to avoid import issues with global onnx package
